move http server code to a server package
