+++ .gitignore
@@ -1,3 +1,4 @@
 *.mod
 *.yaml
 build
+.env
