+++ main_node/.env
@@ -1,2 +1,2 @@
 WALLET_URL=https://spv.money
-ADMIN_XPRIV=your_private_key
+ADMIN_XPRIV=xprv9s21ZrQH143K26NuHinaw24L47fyrs8Vdf52xAHb28SVEW5TEPcZv9qfmMiZhHaAeWT69CfT2prMeyJjxyKVKjFcvZoUKhnYVokCUGEz49o
