+++ main_node/main.go
@@ -3,61 +3,24 @@ package main
 import (
 	"context"
 	"fmt"
-	"log"
-	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
 	"time"
 
-	"github.com/gorilla/mux"
+	"github.com/Du-bem/dominari_scriptum/main_node/server"
 	_ "github.com/joho/godotenv/autoload"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/config"
 )
-
-type AdminData struct {
-	adminAPI *wallet.AdminAPI
-}
-
-func handleUIRequest(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, "UI API data handler is not yet implemented!")
-}
-
-func handleSatelliteRequest(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, "Satellite API data handler is not yet implemented!")
-}
-
-func (a *AdminData) server() {
-	router := mux.NewRouter()
-	router.HandleFunc("/ui", handleUIRequest).Methods("GET")
-	router.HandleFunc("/data", handleSatelliteRequest).Methods("GET")
-
-	srv := &http.Server{
-		Handler:      router,
-		Addr:         "127.0.0.1:8000",
-		WriteTimeout: 15 * time.Second,
-		ReadTimeout:  15 * time.Second,
-	}
-
-	fmt.Println("Running the Server on: ", srv.Addr)
-	log.Fatal(srv.ListenAndServe())
-}
 
 func run(_ context.Context, quit chan error) {
 	fmt.Println("Running the Admin API with with privated Key!")
-	adminAPIData, err := wallet.NewAdminAPIWithXPriv(
-		config.New(config.WithAddr(os.Getenv("WALLET_URL"))), os.Getenv("ADMIN_XPRIV"),
-	)
+	admin, err := server.NewAdminAPI()
 	if err != nil {
 		quit <- err
 		return
 	}
 
-	admin := &AdminData{adminAPI: adminAPIData}
-
-	go admin.server() // Run the http server in a goroutine
+	go admin.RunServer() // Run the http server in a goroutine
 }
 
 func main() {
@@ -95,4 +58,6 @@ func main() {
 	run(ctx, quit)
 
 	<-ctx.Done() // Exit the goroutine once the context is cancelled
+
+	fmt.Println("!!!!! System shutting down !!!!!")
 }
